+++ packages/root/src/types/hickory.ts
@@ -5,6 +5,7 @@ export interface History {
   location: SessionLocation;
   toHref(to: AnyLocation): string;
   respondWith(fn: ResponseHandler): void;
+  cancel(): void;
   destroy(): void;
   navigate(to: ToArgument, navType?: NavType): void;
   go(num?: number): void;