+++ packages/root/types/types/index.d.ts
@@ -1,6 +1,6 @@
 export { History } from "./hickory";
 export { LocationComponents, PartialLocation, SessionLocation, AnyLocation, Location } from "./location";
-export { ToArgument, Action, NavType, PendingNavigation, ResponseHandler } from "./navigation";
+export { ToArgument, Action, NavType, PendingNavigation, ResponseHandler, CancelNavigation } from "./navigation";
 export { KeyFns } from "./keyGenerator";
 export { LocationUtilOptions, QueryFunctions, RawPathname, LocationUtils } from "./locationUtils";
 export { NavigationInfo, ConfirmationFunction, ConfirmationMethods, BlockingHistory } from "./navigationConfirmation";