+++ packages/hash/CHANGELOG.md
@@ -1,3 +1,8 @@
+## Next
+
+- Add `history.cancel` function for cancelling the active navigation.
+- The history object will automatically cancel the pending navigation when another navigation happens.
+
 ## 2.0.0-alpha.3
 
 - Remove generic `query` type.